refactor(dashboard): extract reducer and actions to own file
